+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - New #177: Add `hasTopicsEnabled` field to `User` type.
 - New #177: Add `isNameImplicit` field to `ForumTopic` and `ForumTopicCreated` types.
+- New #177: Add `giftUpgradeSent` field to `Message` type.
 - New #177: Add `SendMessageDraft`, `getUserGifts` and `getChatGifts` methods.
 - Chg #177: Replace `lastResaleStarCount` field with `lastResaleCurrency` and `lastResaleAmount` fields in
   `UniqueGiftInfo` type.